chore: remove --harmony-async-await flag
With node 7.6.0 we have native async/await. The containers run 7.10.1
(you can verify this by running `docker run mhart/alpine-node:7 node --version`),
so there is no need to have this extra flag, which will cause the thing
to break on current node versions (tested with 8.9.4).

In future, we probably should switch to the LTS cycle, which works fine
on all of the GK code that I've touched so far.